模型接口测试结束

+++ digital-service/src/main/java/com/digital/model/request/threejs/VisualReq.java
@@ -16,5 +16,5 @@ public class VisualReq implements Serializable {
     private String leftAstigmatism;
     private String rightAstigmatism;
     private String colorVision;
-    private Integer kidId;
+    private Long kidId;
 }
